Initiate changes in Configurations:     - Introduce defaults for some parameters.     - Remove unnecessary panic()
