+++ Cargo.toml
@@ -1,4 +1,4 @@
-workspace = { members = ["vrrp_packet"] }
+workspace = { members = [ "tun-tap", "vrrp_packet"] }
 [package]
 name = "failover"
 version = "0.2.4"
@@ -17,8 +17,8 @@ rand = "0.8.5"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"
 simple_logger = "4.3.3"
-tun-tap = "0.1.4"
 getopts = "0.2"
+libc = "0.2.153"
 
 [lib]
 doctest = false
