+++ src/general.rs
@@ -1,14 +1,17 @@
 
-use std::{env, error::Error, fs::{self, File}, io::{BufReader, Write}, path::Path, process::Command, str::FromStr};
-use crate::{config::{CliConfig, FileConfig, VrrpConfig}, error::OptError, router::VirtualRouter, state_machine::VirtualRouterMachine};
-use getopts::Options;
+use std::{process::Command, str::FromStr};
+use crate::{config::VrrpConfig, router::VirtualRouter, state_machine::VirtualRouterMachine};
 use pnet::datalink::{self, Channel, DataLinkReceiver, DataLinkSender, NetworkInterface};
 use ipnet::Ipv4Net;
+use rand::{distributions::Alphanumeric, Rng};
 
 pub(crate) fn get_interface(name: &str) -> NetworkInterface {
     let interface_names_match = |iface: &NetworkInterface| iface.name == name;
     let interfaces = datalink::linux::interfaces();
+
+    // check if interface name exists, if not create it
     interfaces.into_iter().find(interface_names_match).unwrap()
+    
 }
 
 pub(crate) fn create_datalink_channel(interface: &NetworkInterface)  -> (Box<dyn DataLinkSender>, Box<dyn DataLinkReceiver>){
@@ -22,20 +25,6 @@ pub(crate) fn create_datalink_channel(interface: &NetworkInterface)  -> (Box<dyn
     }
 }
 
-const DEFAULT_JSON_CONFIG: &'static [u8; 201] = b"
-{
-    \"name\": \"VR_1\",
-    \"vrid\": 51,
-    \"interface_name\": \"wlo1\",
-    \"ip_addresses\": [
-        \"192.168.100.100/24\"
-    ],
-    \"priority\": 101,
-    \"advert_interval\": 1,
-    \"preempt_mode\": true
-}
-";
-
 // takes the configs that have been received and converts them 
 // into a virtual router instance. 
 pub fn config_to_vr(conf: VrrpConfig) -> VirtualRouter {
@@ -47,17 +36,21 @@ pub fn config_to_vr(conf: VrrpConfig) -> VirtualRouter {
     let master_down_interval: f32 = (3_f32 * conf.advert_interval() as f32) + skew_time;
     
     let mut ips: Vec<Ipv4Net> = vec![];
-    for ip_config in &conf.ip_addresses() {
+
+    if conf.ip_addresses().len() > 20 {
+        log::warn!("({})  More than 20 IP addresses(max for VRRP) have been configured. Only first 20 addresses will be considered. ", conf.name());
+    }
+
+    let addresses = if conf.ip_addresses().len() <= 20 { conf.ip_addresses() } else { conf.ip_addresses()[0..20].to_vec() }; 
+    for ip_config in addresses.iter() {
         match Ipv4Net::from_str(ip_config) {
             Ok(ip_addr) => ips.push(ip_addr),
-            Err(err) => {
-                log::error!("Address '{:?}' not in the correct format", &ip_config);
-                panic!("Error: {err}");
+            Err(_) => {
+                log::error!("({}) SKIPPING: Configured IP address '{:?}' not in the correct format. ", ip_config, conf.name());
             }
         }
     }
     
-    log::info!("({}) Setting up.", conf.name());
     let vr = VirtualRouter {
         name: conf.name().clone(),
         vrid: conf.vrid(),
@@ -76,215 +69,6 @@ pub fn config_to_vr(conf: VrrpConfig) -> VirtualRouter {
 }
 
 
-pub fn parse_cli_opts(args: &[String]) -> Result<VrrpConfig, OptError>{
-    let mut opts = Options::new();
-
-    opts.optflag("H", "help", "display help information");
-    opts.optflag("C", "cli", "use the cli config option");
-    
-    opts.optopt(
-        "A", 
-        "action", 
-        "action that will be done to the addresses on the interface configured. Default is 'run'", 
-        "(--action teardown / --action run)");
-
-    // name
-    opts.optopt(
-        "n", 
-        "name", 
-        "name of the virtual router instance.",
-    "(--name VR_1)");
-    
-    // vrid 
-    opts.optopt(
-        "v", 
-        "vrid", 
-        "The Virtual Router ID of the instance. In the range of 1-255", 
-        "(--vrid 51)");
-
-    // ip addresses 
-    opts.optmulti(
-    "I", 
-    "ip-address", 
-    "An Ip address that is associated with the virtual router instance", 
-    "(--ip-address 192.168.100.5/24)");
-
-    // interface name
-    opts.optopt(
-        "i", 
-        "iface", 
-        "The interfaece that the virtual IP(s) will be attached to.", 
-        "(--iface eth0)");
-    
-    // priority
-    opts.optopt(
-        "p", 
-        "priority", 
-        "priority of the virutal router in the VRRP network group. In the range 1-44", 
-        "(--priority 100)");
-    
-    opts.optopt(
-        "a", 
-        "adv-interval", 
-        "When in master, the interval when ADVERTISEMENTS should be carried across", 
-    "(--adv-interval 2)");
-
-    opts.optopt(
-        "P", 
-        "preempt-mode", 
-        "Controls whether a higher priority Backup router preempts a lower priority Master.", 
-    "(--preempt-mode false)");
-
-    opts.optopt(
-        "f", 
-        "file", 
-        "the json file with the necessary configurations. By default will be looked for at: '{CURRENT_PATH}/vrrp-config.json'", 
-    "(--file FILENAME)");
-
-
-    // if it is the help request
-    if args[1..].contains(&"--help".to_string()) {
-        let help_format = "
-        Failover Usage:
-            # running failover, we take configs either from a json file or from the cli   
-            CONFIG
-            ======
-
-            FILE CONFIG MODE
-            ----------------
-            ./failover --file custom-vrrp-config.json
-
-            CLI CONFIG MODE 
-            ---------------
-            ./failover --cli --iface wlo1 --priority 101 --adv-interval 1 --preempt-mode false
-
-            DEFAULT
-            -------
-            ./failover 
-            # if neither 'file' not 'cli' is specified, failover chooses 'file' by default. 
-            # The file that is used for configs is '{CURRENT_PATH}/vrrp-config.json' in the same 
-            # directory where failover is being run from (TODO: to change this to relevant config file in /etc directory).
-
-            ACTIONS
-            =======
-            # Two actions can be run: 'teardown' or 'run'. 
-            # 'run' is default if no actions are specified. 
-            ./failover --teardown
-            
-            # can also be called without --run
-            ./failover --run 
-            ./failover --teardown
-
-        ".to_string();
-        
-        println!("{}", opts.usage(&help_format));
-        std::process::exit(0);
-    }
-
-    let matches = match opts.parse(&args[1..]) {
-        Ok(m) => m,
-        Err(err) => return Result::Err(OptError(err.to_string()))
-    };
-
-    if matches.opt_str("cli").is_some() {
-        
-        let cli_config = CliConfig {
-            name: match matches.opt_str("name") {
-                Some(x) => x,
-                None => return Result::Err(OptError("instance name '--name' is a mandatory field".into()))
-            },
-            vrid: match matches.opt_str("vrid") {
-                Some(x) => x.parse::<u8>().unwrap(),
-                None => return Result::Err(OptError("VRID '--vrid' is a mandatory field".into()))
-            },
-            interface_name: match matches.opt_str("iface") {
-                Some(x) => x,
-                None => return Result::Err(OptError("interface name '--iface' is a mandatory field".into()))
-            },
-            ip_addresses: matches.opt_strs("ip-address"),
-            priority: match matches.opt_str("priority") {
-                Some(x) => x.parse::<u8>().unwrap(),
-                None => CliConfig::default().priority
-            },
-            advert_interval: match matches.opt_str("adv-interval") {
-                Some(x) => x.parse::<u8>().unwrap(),
-                None => CliConfig::default().advert_interval
-            },
-            preempt_mode: match matches.opt_str("preempt-mode") {
-                Some (x) => x.parse::<bool>().unwrap(),
-                None => CliConfig::default().preempt_mode
-            }
-        };
-
-        match matches.opt_str("action") {
-            Some(x) => {
-
-                if ["teardown"].contains(&x.to_lowercase().as_str()){
-                    virtual_address_action("delete", &cli_config.ip_addresses, &cli_config.interface_name);
-                    std::process::exit(0);
-                } else {
-                    return Result::Err(OptError("--action has to be ether 'run' or 'teardown' . If none is specified, run will be default.".into()));
-                }
-            }
-            None => {
-                // should have 'run' as action by default if nothing is specified.  
-            }
-        }
-
-
-        Ok(VrrpConfig::Cli(cli_config))
-    } else {
-
-        let filename = if matches.opt_str("file").is_some() { 
-            matches.opt_str("file").unwrap() 
-        } else {
-
-            // if app is running via snap, the SNAP_COMMON environment 
-            // variable will be used as the config directory
-            let directory = match env::var("SNAP_COMMON") {
-                Ok(path) => path + "/",
-                Err(_) => {
-                    let _ = fs::create_dir_all("/etc/failover/");
-                    "/etc/failover/".to_string()
-                }
-            };
-
-            let file_path = &format!("{}vrrp-config.json", directory);
-
-            if !Path::new(file_path).exists() {
-                let mut file = File::create(file_path).unwrap();
-                let _ = file.write_all(DEFAULT_JSON_CONFIG);
-            }
-            file_path.to_string()
-        };
-        let file_config = match read_config_from_json_file(&filename) {
-            Ok(config) => VrrpConfig::File(config),
-            Err(err) => {
-                log::error!("{err}");
-                return  Result::Err(OptError(format!("Problem Parsing file {}", &filename)))
-            }
-        };
-
-        match matches.opt_str("action") {
-            Some (x) => {
-                if ["setup", "teardown"].contains(&x.to_lowercase().as_str()){
-                    let action = if x.to_lowercase().as_str() == "setup" { "add" } else { "delete" };
-                    virtual_address_action(action, &file_config.ip_addresses(), &file_config.interface_name());
-                    std::process::exit(0);
-                } else if !["run"].contains(&x.to_lowercase().as_str()) {
-                    return Result::Err(OptError("--action has to be ether 'setup', 'teardown' or 'run' ".into()));
-                } else {
-                    Ok(file_config)
-                }
-            }
-            
-            None => {
-                Ok(file_config)
-            }
-        }
-    } 
-}
-
 pub(crate) fn virtual_address_action(action: &str, addresses: &[String], interface_name: &str)
 {
     for addr in addresses {
@@ -296,12 +80,10 @@ pub(crate) fn virtual_address_action(action: &str, addresses: &[String], interfa
 }
 
 
-
-fn read_config_from_json_file<P: AsRef<Path>>(path: P) -> Result<FileConfig, Box<dyn Error>> 
-{
-    log::info!("Reading from config file {:?}", path.as_ref().as_os_str());
-    let file = File::open(path)?;
-    let reader = BufReader::new(file);
-    let u = serde_json::from_reader(reader)?;
-    Ok(u)
+pub(crate) fn random_string(length: usize) -> String {
+    rand::thread_rng()
+        .sample_iter(&Alphanumeric)
+        .take(length)
+        .map(char::from)
+        .collect()
 }
