+++ src/pkt/generators.rs
@@ -65,10 +65,6 @@ impl MutablePktGenerator
         vrrp_pkt.set_auth_data2(0);
         vrrp_pkt.set_ip_addresses(&addresses);
         
-        if vrrp_pkt.get_ip_addresses().len() > 20 {
-            log::error!("VRRP packet cannot have more than 20 IP addresses. Fix in your configs.");
-            panic!("VRRP configuration VRID={} has more than 20 IP addresses", vrrp_pkt.get_vrid());
-        }
         vrrp_pkt
     }
